src/hr_mcp_functions/comms_mcp/requirements.txt
azure-functions
fastmcp
pydantic
pydantic-settings
azure-identity
httpx
